+++ CHANGELOG.rst
@@ -10,6 +10,14 @@ Note that the issue IDs here refer to ones in the private CUBI GitLab.
 Unreleased
 ==========
 
+Added
+-----
+
+- **Adminalerts**
+    - UI documentation (#865)
+- **Siteinfo**
+    - UI documentation (#865)
+
 Changed
 -------
 
@@ -18,6 +26,8 @@ Changed
     - Upgrade to python-ldap v3.4.0 (#871)
 - **Projectroles**
     - HTTP 403 raised instead of 400 if project type disallowed by API view (#872)
+    - Update role list media rules (#863)
+    - Add line break for custom email footer (#864)
 
 Fixed
 -----