+++ CHANGELOG.rst
@@ -7,6 +7,38 @@ Changelog for the **SODAR Core** Django app package. Loosely follows the
 Note that the issue IDs here refer to ones in the private CUBI GitLab.
 
 
+v0.10.7 (2021-12-14)
+====================
+
+Added
+-----
+
+- **Adminalerts**
+    - UI documentation (#865)
+- **Siteinfo**
+    - UI documentation (#865)
+
+Changed
+-------
+
+- **General**
+    - Upgrade minimum Django version to v3.2.10 (#869)
+    - Upgrade to python-ldap v3.4.0 (#871)
+- **Projectroles**
+    - HTTP 403 raised instead of 400 if project type disallowed by API view (#872)
+    - Update role list media rules (#863)
+    - Add line break for custom email footer (#864)
+
+Fixed
+-----
+
+- **Projectroles**
+    - ``ManagementCommandLogger`` crash by unset ``LOGGING_LEVEL`` (#862)
+    - ``highlight_search_term()`` crash on invalid term input (#867)
+    - Search bar allowing invalid input (#868)
+    - Wrong project type displayed in project type restriction API response (#872)
+
+
 v0.10.6 (2021-11-19)
 ====================